Initial commit
I was just going for an experiment here, there is much more to be
implemented yet.
Although I love the idea of normalizing my data (as that makes updating
the dataset very trivial), it has a serious drawback. Accessing that
same dataset becomes very cumbersome.

This library attemps to do just that. Fix the way we access our data
after the dataset has been normalized.